fix: move application utilities under `application` module
