+++ rusaint/src/utils/mod.rs
@@ -18,6 +18,3 @@ pub(crate) fn default_header() -> HeaderMap {
     headers.insert(CONNECTION, "keep-alive".parse().unwrap());
     headers
 }
-
-pub(crate) mod de_with;
-pub(crate) mod input_field;
